+++ test/LionWeb.Protocol.Delta.Test/Pipe/TimeDependent/TimeDependentReplicatorTestsBase.cs
@@ -314,7 +314,7 @@ public abstract class TimeDependentReplicatorTestsBase  : ReplicatorTestsBase
     }
 
     [TestMethod]
-    public void ChildMovedAndReplacedInSameContainment()
+    public void ChildMovedAndReplacedInSameContainmentInSameParent()
     {
         var grandChild = new LinkTestConcept("grandChild");
         var child = new LinkTestConcept("child") { Containment_0_1 = grandChild };
@@ -328,9 +328,9 @@ public abstract class TimeDependentReplicatorTestsBase  : ReplicatorTestsBase
         child.Containment_0_1 = null;
         var composite = _compositor.Pop(true);
         
-        Assert.IsInstanceOfType<ChildMovedAndReplacedInSameContainmentNotification>(composite.Parts[^1]);
+        Assert.IsInstanceOfType<ChildMovedAndReplacedInSameContainmentInSameParentNotification>(composite.Parts[^1]);
 
-        var deleted = (ChildMovedAndReplacedInSameContainmentNotification)composite.Parts[^1];
+        var deleted = (ChildMovedAndReplacedInSameContainmentInSameParentNotification)composite.Parts[^1];
         CollectionAssert.AreEqual(new List<IReadableNode>{child, grandChild}, deleted.DeletedNodes.ToList());
 
         AssertPartitionsEqual();